Flatten directory structure
