+++ packages/grid/apps/domain/src/main/core/database/bin_obj.py
@@ -1,13 +1,10 @@
 # third party
 from syft import deserialize
 from syft import serialize
-from syft.proto.lib.pandas.frame_pb2 import PandasDataFrame as PandasDataFrame_PB
-from syft.proto.lib.torch.module_pb2 import Module as Module_PB
-from syft.proto.lib.torch.tensor_pb2 import TensorProto as TensorProto_PB
 
 # grid relative
-from .. import BaseModel
-from .. import db
+from . import BaseModel
+from . import db
 
 
 class BinObject(BaseModel):
